An  analysis done with Microsoft Excel.
In this project, data from a coffee shop was cleaned and analyzed using Microsoft Excel. I made use of IF functions, VLOOK UP and XLOOK UP, while working on the dataset.
Insights on the top customers, sales for each coffee type over the given time frame and sales for each country were drawn and visualized.